Add files via upload
This is the first time I've used the tuple format when grouping data. The syntax is much cleaner than the dictionary method, and crucially you're able to call information from the same column multiple times if looking to perform different operations on that column. For example when looking to both work out the total as well as the mean.